# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/produk": {
            "get": {
                "tags": [
                    "produk"
                ],
                "summary": "Menampilkan semua produk",
                "operationId": "82be1cfb35a47c9a1908501faad2fe91",
                "responses": {
                    "200": {
                        "description": "List of Produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "produk"
                ],
                "summary": "Membuat produk baru",
                "operationId": "f5c50d3af3d9a2689821c4c6ea313cc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_produk",
                                    "deskripsi",
                                    "harga",
                                    "stok",
                                    "id_kategori"
                                ],
                                "properties": {
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "id_kategori": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/produk/{id}": {
            "get": {
                "tags": [
                    "produk"
                ],
                "summary": "Menampilkan detail produk berdasarkan ID",
                "operationId": "8d795c6c96c869876803d9b62ed02820",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk found"
                    },
                    "404": {
                        "description": "Produk not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "produk"
                ],
                "summary": "Mengupdate data produk",
                "operationId": "1b02fbf6cf518c68d4b3bd597d9ca2a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "id_kategori": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Produk not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "produk"
                ],
                "summary": "Menghapus produk",
                "operationId": "1c8c6cacbfdbc48c6f718a3b6832b279",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk deleted successfully"
                    },
                    "404": {
                        "description": "Produk not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}